+++ 5-es-modules/1-import-and-export/exercises/exercise-02/user.js
@@ -1,5 +1,3 @@
-// #todo
-
 debugger;
 
-export const user = _;
+export const user = 'fendy';
