+++ 5-es-modules/1-import-and-export/exercises/exercise-04/say-hi.js
@@ -1,5 +1,5 @@
-// #todo
-
 debugger;
 
-const sayHi = __;
+export const sayHi = (name) => {
+  return `Hello, ${name}!`;
+};
