+++ 5-es-modules/2-spec-files/exercises/3-pass-tests/type-of-value.js
@@ -1,8 +1,6 @@
-// #todo
-
 /**
  * returns the type of any value
  * @param {any} value - any javascript value
  * @returns {string} the value's type
  */
-__;
+export const typeOfValue = (value) => typeof value;
